+++ docs/master/sidebar.js
@@ -22,6 +22,8 @@ module.exports = [
         children: [
             ['eloquent/getting-started', 'Getting Started'],
             'eloquent/relationships',
+            'eloquent/polymorphic-relationships',
+            'eloquent/soft-deleting',
             'eloquent/nested-mutations',
         ]
     },