+++ src/Schema/Directives/CreateDirective.php
@@ -10,8 +10,9 @@ use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Execution\MutationExecutor;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class CreateDirective extends BaseDirective implements FieldResolver
+class CreateDirective extends BaseDirective implements FieldResolver, DefinedDirective
 {
     /**
      * @var \Illuminate\Database\DatabaseManager
@@ -35,6 +36,22 @@ class CreateDirective extends BaseDirective implements FieldResolver
     public function name(): string
     {
         return 'create';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Create a new Eloquent model with the given arguments.
+"""
+directive @create(  
+  """
+  Specify the class name of the model to use.
+  This is only needed when the default model resolution does not work.
+  """
+  model: String
+) on FIELD_DEFINITION
+SDL;
     }
 
     /**