+++ src/Support/Contracts/DefinedDirective.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Nuwave\Lighthouse\Support\Contracts;
+
+interface DefinedDirective
+{
+    /**
+     * SDL definition of the directive.
+     *
+     * @return string
+     */
+    public static function definition();
+}