+++ CHANGELOG.md
@@ -5,12 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.1.1...master)
+## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.2.1...master)
 
 ### Added
 
+- Add `@softDeletes` and `@trashed` directives to enable
+  filtering soft deleted models https://github.com/nuwave/lighthouse/pull/937
+
+## [4.2.1](https://github.com/nuwave/lighthouse/compare/v4.2.0...v4.2.1)
+
+### Fixed
+
+- Actually use the specified `edgeType` in Relay style connections https://github.com/nuwave/lighthouse/pull/939
+
+## [4.2.0](https://github.com/nuwave/lighthouse/compare/v4.1.1...v4.2.0)
+
+### Added
+
+- Add `@morphOne` directive for polymorphic one-to-one relationships https://github.com/nuwave/lighthouse/pull/944
 - Add `@morphTo` directive for polymorphic one-to-one relationships https://github.com/nuwave/lighthouse/pull/921
+- Add `@morphMany` directive for polymorphic one-to-many relationships https://github.com/nuwave/lighthouse/pull/944
 - Support Laravel `^6.0` https://github.com/nuwave/lighthouse/pull/926
+- Add command `lighthouse:ide-helper` for generating a definition file with all schema directives https://github.com/nuwave/lighthouse/pull/933
 
 ## [4.1.1](https://github.com/nuwave/lighthouse/compare/v4.1.0...v4.1.1)